requirements.txt
kafka-python
transformers
torch
cloudevents
kafka-python
transformers
cloudevents
requests